backend/requirements.txt
fastapi == 0.109.2

uvicorn == 0.27.1

motor == 3.3.2

langchain_openai
langchain_community
python-dotenv
pydantic[email]
langchain
chromadb
PyJWT